Module zerocaf::backend::u64::constants

This module contains backend-specific constant values as the 64-bit limbs of curve constants.

Constants

BASEPOINT
Holds the value of the Curve basepoint, which has been constructed from taking y-coodrinate = 3/5 (mod l).

FIELD_L
FIELD_L is the order of the Prime field for Sonny, n this case it is equivalent to 2^252 + 27742317777372353535851937790883648493`

INVERSE_MOD_TWO
INVERSE_MOD_TWO = 1/2 (mod l).

INV_RR
Montgomery modulus defined for FieldElements on inverse() functions scope. It is used for the Montgomery Mul operation that takes place on the `Inversion

INV_SQRT_A_MINUS_D
INV_SQRT_A_MINUS_D = 482283834104289360917429750399313974390948281833312135312952165682596457149.

L
L is the order of base point for Sonny, in this case it is equivalent to 2^249 + 14490550575682688738086195780655237219

LFACTOR
Scalar-LFACTOR is the value that satisfies the equation: L * LFACTOR = -1 (mod 2^52) In this case, LFACTOR is the one used for the Montgomery Reduction algorithm, implemented on Scalar Arithmetics module.

LFACTOR_FIELD
FieldElement-LFACTOR is the value that satisfies the equation: L * LFACTOR = -1 (mod 2^52) In this case, LFACTOR is the one used for the Montgomery Reduction algorithm, implemented on FieldElement Arithmetics module.

MINUS_ONE_HALF
MINUS_ONE_HALF = (-1/2) (mod l).

RISTRETTO_BASEPOINT
Ristretto Basepoint.

RR
Montgomery modulus defined for Scalar arithmetics, R^2 = (2^260)^2 % L

RR_FIELD
Montgomery modulus defined for FieldElement arithmetics, R^2 = (2^260)^2 % FIELD_L

SCALAR_INVERSE_MOD_TWO
SCALAR_INVERSE_MOD_TWO = 1/2 (mod l).

SQRT_AD_MINUS_ONE
SQRT_AD_MINUS_ONE = .

SQRT_MINUS_ONE
Holds the value of one of both sqrt(-1 (mod p)) values. SQRT_MINUS_ONE = 3034649101460298094273452163494570791663566989388331537498831373842135895065`.

Statics

EDWARDS_A
Edwards a variable value = -1 (mod l) equals: 7237005577332262213973186563042994240857116359379907606001950938285454250988 where l = Prime of the field = 2^252 + 27742317777372353535851937790883648493

EDWARDS_D
Edwards d variable value = -126296/126297 (mod l) equals: 951605751702391019481481818669129158712512026257330939079110344917983315091 where l = Prime of the field = 2^252 + 27742317777372353535851937790883648493

INV_SQRT_A
(+)1/SQRT(a) (mod l) equals: 4202356475871964119699734399548423449193549369991576068503119564443318355924.

MINUS_SQRT_A
(-)SQRT(a) (mod l) equals: 4202356475871964119699734399548423449193549369991576068503119564443318355924.