Struct zerocaf::backend::u64::field::FieldElement

pub struct FieldElement(pub [u64; 5]);

A FieldElement represents an element into the field 2^252 + 27742317777372353535851937790883648493

In the 64-bit backend implementation, the FieldElement is represented in radix 2^52

Methods

impl FieldElement

pub fn zero() -> FieldElement

Construct zero.

pub fn one() -> FieldElement

Construct one.

pub fn minus_one() -> FieldElement

Construct -1 (mod l).

pub fn is_even(self) -> bool

Evaluate if a FieldElement is even or not.

pub fn half(&self) -> FieldElement

Give the half of the FieldElement value (mod l). This function SHOULD ONLY be used with even FieldElements otherways, can produce erroneus results.

pub fn plus_p_and_half(&self) -> FieldElement

Performs the operation ((a + constants::FIELD_L) >> 2) % l). This function SHOULD only be used on the Kalinski's modular inverse algorithm, since it's the only way we have to addlto aFieldElement` without obtaining the same number.

On Kalinski's PhaseII, this function allows us to trick the addition and be able to divide odd numbers by 2.

pub fn from_bytes(bytes: &[u8; 32]) -> Self

Load a FieldElement from the low 253b bits of a 256-bit input. So Little Endian representation in bytes of a FieldElement.

pub fn to_bytes(self) -> [u8; 32]

Serialize this FieldElement to a 32-byte array. The encoding is canonical.

pub fn two_pow_k(exp: &u64) -> FieldElement

Given a k: u64, compute 2^k giving the resulting result as a FieldElement. Note that the input must be between the range => 0..260.

NOTE: Usually, we will say 253, but since on some operations as inversion we need to exponenciate to greater values, we set the max on the Montgomery modulo so 260.

pub fn square(&self) -> FieldElement

Compute self^2 (mod l).

pub fn mul_internal(a: &FieldElement, b: &FieldElement) -> [u128; 9]

Compute a * b with the function multiplying helper

pub fn montgomery_reduce(limbs: &[u128; 9]) -> FieldElement

Compute limbs/R (mod l), where R is the Montgomery modulus 2^260

pub fn montgomery_mul(a: &FieldElement, b: &FieldElement) -> FieldElement

Compute (a * b) / R (mod l), where R is the Montgomery modulus 2^253

pub fn to_montgomery(&self) -> FieldElement

Puts a Scalar into Montgomery form, i.e. computes a*R (mod l)

pub fn from_montgomery(&self) -> FieldElement

Takes a FieldElement out of Montgomery form, i.e. computes a/R (mod l)

pub fn kalinski_inverse(a: &FieldElement) -> FieldElement

Compute a^-1 (mod l) using the the Kalinski implementation of the Montgomery Modular Inverse algorithm. B. S. Kaliski Jr. - The Montgomery inverse and its applica-tions. IEEE Transactions on Computers, 44(8):1064–1065, August-1995

pub fn savas_koc_inverse(a: &FieldElement) -> FieldElement

Compute a^-1 (mod l) using the the Savas & Koç modular inverse algorithm. It's an optimization of the Kalinski modular inversion algorithm that extends the Binary GCD algorithm to perform the modular inverse operation.

The PhaseII it's substituded by 1 or 2 Montgomery Multiplications, what makes the second part compute in almost ConstTime.

SPECIAL ISSUE ON MONTGOMERY ARITHMETIC. Montgomery inversion - Erkay Sava ̧s & Çetin Kaya Koç J Cryptogr Eng (2018) 8:201–210 https://doi.org/10.1007/s13389-017-0161-x