Module zerocaf::backend::u64::scalar

Arithmetic mod 2^249 + 14490550575682688738086195780655237219 with five 52-bit unsigned limbs represented in radix 2^52.

//! The basic modular operations have been taken from the curve25519-dalek repository and refactored to work for the Sonny sub-group field.

Structs

Scalar
The Scalar struct represents an Scalar over the modulo 2^249 + 14490550575682688738086195780655237219 as 5 52-bit limbs represented in radix 2^52.