Struct zerocaf::backend::u64::scalar::Scalar

pub struct Scalar(pub [u64; 5]);

The Scalar struct represents an Scalar over the modulo 2^249 + 14490550575682688738086195780655237219 as 5 52-bit limbs represented in radix 2^52.

Methods

impl Scalar

pub const fn zero() -> Scalar

Return a Scalar with value = 0.

pub const fn one() -> Scalar

Return a Scalar with value = 1.

pub const fn minus_one() -> Scalar

Return a Scalar with value = -1 (mod l).

pub fn is_even(self) -> bool

Evaluate if a Scalar is even or not.

pub fn from_bytes(bytes: &[u8; 32]) -> Scalar

Unpack a 32 byte / 256 bit Scalar into 5 52-bit limbs.

pub fn from_bytes_wide(_bytes: &[u8; 64]) -> Scalar

Reduce a 64 byte / 512 bit scalar mod l

pub fn to_bytes(&self) -> [u8; 32]

Pack the limbs of this Scalar into 32 bytes

pub fn two_pow_k(exp: u64) -> Scalar

Given a k: u64, compute 2^k giving the resulting result as a Scalar.

See that the input must be between the range => 0..250.

NOTE: This function implements an assert! statement that checks the correctness of the exponent provided as param.