Module zerocaf::constants

Contains the curve-constants needed by different algorithm implementations.

Re-exports

pub use crate::backend::u64::constants::*;

Constants

BASEPOINT_COMPRESSED
Holds the value of the Curve basepoint, which has been constructed from taking y-coodrinate = 3/5 (mod l). The positive sign is choosen for it, so we leave it on it's cannonical bytes encoding.

RISTRETTO_BASEPOINT_COMPRESSED
Ristretto Basepoint on compressed format.