Struct zerocaf::edwards::AffinePoint

pub struct AffinePoint {
    pub X: FieldElement,
    pub Y: FieldElement,
}

An AffinePoint represents a point on the Sonny Curve expressed over the Twisted Edwards Affine Coordinates also known as cartesian coordinates: (X, Y).

This Twisted Edwards coordinates are ONLY implemented for equalty testing, since all of the Point operations defined over them are much slower than the same ones defined over Extended or Projective coordinates.

Fields
X: FieldElementY: FieldElement