Module zerocaf::montgomery

Implementation that provides support for Montgomery Points over the Sonnycurve.

A MontgomeryPoint is represented as the u-coordinate of itself in LE bytes-format.

Structs

MontgomeryPoint
Holds the u-coordinate of a point on the Montgomery form of Doppio-curve or its twist.