Struct zerocaf::ristretto::RistrettoPoint

pub struct RistrettoPoint(pub EdwardsPoint);

Methods

impl RistrettoPoint

pub fn compress(&self) -> CompressedRistretto

Encode a Ristretto point represented by the point (X:Y:Z:T) in extended coordinates.

pub fn elligator_ristretto_flavor(r_0: &FieldElement) -> RistrettoPoint

Computes the Ristretto Elligator map. This gets a RistrettoPoint from a given `FieldElement´.

pub fn from_uniform_bytes(bytes: &[u8; 64]) -> RistrettoPoint

Construct a RistrettoPoint from 64 bytes of data.

If the input bytes are uniformly distributed, the resulting point will be uniformly distributed over the group, and its discrete log with respect to other points should be unknown.

Implementation

This function splits the input array into two 32-byte halves, takes the low 255 bits of each half mod p, applies the Ristretto-flavored Elligator map to each, and adds the results.

This function is taken from the Ristretto255 implementation found in curve25519-dalek

pub fn new_random_point<T: Rng + CryptoRng>(rand: &mut T) -> RistrettoPoint

Generate a random RistrettoPoint from a 64-byte array generated with user-provided rng.

The provided rng has to implement: Rng + CryptoRng.

This function uses the elligator hash map twice, once for [0..31] & another for [32..64] giving a uniformly distributed random value.

This implementation follows the idea pointed on the random point generation used in curve25519-dalek.