Type Definition zerocaf::scalar::Scalar

type Scalar = Scalar;

A Scalar represents an element of the field generated by the prime of the sub-group: 2^249 - 15145038707218910765482344729778085401.

This is a type alias for one of the Scalar types in the backend module.