Module zerocaf::traits::ops

Traits

Double
Trait that represents the Point doubling operation for any type of Point that is used on the lib.

Half
Trait that represents the /2 operation for any kind of element on the library.

InvSqrt

ModSqrt

Pow
Trait that represents the modular exponentiation operation, ie.a ^ b (mod l), for any kind of element on the library (except points).

SqrtRatioI

Square
Trait that represents the ^2 operation for any kind of element on the library.