Trait zerocaf::traits::ops::Double

pub trait Double {
    type Output;
#[must_use]
    fn double(self) -> Self::Output;
}

Trait that represents the Point doubling operation for any type of Point that is used on the lib.

This trait is implemented following the rules that mandate over the Type that is being implemented.

Associated Types

type Output

Loading content...
Required methods

#[must_use] fn double(self) -> Self::Output

Performs the point-doubling operation over the coordinates which this trait has been implemented for.

Loading content...