Trait zerocaf::traits::ops::InvSqrt

pub trait InvSqrt {
    type Output;
#[must_use]
    fn inv_sqrt(self) -> Self::Output;
}

Associated Types

type Output

Loading content...
Required methods

#[must_use] fn inv_sqrt(self) -> Self::Output

Performs the Inverse Square root of a given value.

This operation returns always the positive result of the modular sqrt, understanding positive as the definition that appears on the Decaf paper: 0 < result < (P - 1)/2.

Loading content...