Trait zerocaf::traits::ops::ModSqrt

pub trait ModSqrt {
    type Output;
#[must_use]
    fn mod_sqrt(self, choice: Choice) -> Self::Output;
}

Associated Types

type Output

Loading content...
Required methods

#[must_use] fn mod_sqrt(self, choice: Choice) -> Self::Output

Performs the modular Square Root operation over a finite field ie. sqrt(x) (mod l).

With the given Choice, the impl is able to provide the result that corresponds to the positive or negative sign choosen.

Returns

Some(symb_choosen_result) if the input is a QR for the prime modulo. Otherways it returns None

Loading content...