Trait zerocaf::traits::ops::Pow

pub trait Pow<T> {
    type Output;
#[must_use]
    fn pow(self, exp: T) -> Self::Output;
}

Trait that represents the modular exponentiation operation, ie.a ^ b (mod l), for any kind of element on the library (except points).

This trait is implemented following the rules that mandate over the Type that is being implemented.

Associated Types

type Output

Loading content...
Required methods

#[must_use] fn pow(self, exp: T) -> Self::Output

Returns a^b (mod l).

Loading content...