Trait zerocaf::traits::ops::Square

pub trait Square {
    type Output;
#[must_use]
    fn square(self) -> Self::Output;
}

Trait that represents the ^2 operation for any kind of element on the library.

This trait is implemented following the rules that mandate over the Type that is being implemented.

Associated Types

type Output

Loading content...
Required methods

#[must_use] fn square(self) -> Self::Output

Returns the square of the input: x^2.

Loading content...