Trait zerocaf::traits::Identity

pub trait Identity {
#[must_use]
    fn identity() -> Self;
}

Gives the Identity element for the type which it has been implemented on.

This trait is implemented following the rules that mandate over the Type that is being implemented.

Required methods

#[must_use] fn identity() -> Self

Returns the identity element for the implemented type, following it's mathematical rules.

Loading content...