Trait zerocaf::traits::ValidityCheck

pub trait ValidityCheck {
#[must_use]
    fn is_valid(&self) -> Choice;
}

This trait pretends to be a verification in ct_time about a point correctness.

This is done through checking that the (X, Y) coordinates of the point are valid and satisfy the curve equation.

Required methods

#[must_use] fn is_valid(&self) -> Choice

Checks the point coordinates agains the curve equation to validate that the point relies on the curve and its valid.

Returns

Choice(0) if the point isn't a valid point.

Choice(1) if the point is valid.

Loading content...