Module zerocaf::backend::u64::constants

This module contains backend-specific constant values, such as the 64-bit limbs of curve constants.

Constants

FIELD_L
FIELD_L is the order of the Prime field for Doppio, i.e. 2^252 + 27742317777372353535851937790883648493`

INV_RR
Montgomery modulus defined for FieldElements on inverse() function's scope. It's used for the Montgomery Mul operation that takes place on the Inversion operation. It's defined as: R^2 = (2^253)^2 % L

L
L is the order of base point for Doppio, i.e. 2^249 - 15145038707218910765482344729778085401

LFACTOR
Scalar-LFACTOR is the value that satisfies the equation: L * LFACTOR = -1 (mod 2^52) In this case, LFACTOR is the one used for the Montgomery Reduction algorithm, implemented on Scalar Arithmetics module.

LFACTOR_FIELD
FieldElement-LFACTOR is the value that satisfies the equation: L * LFACTOR = -1 (mod 2^52) In this case, LFACTOR is the one used for the Montgomery Reduction algorithm, implemented on FieldElement Arithmetics module.

RR
Montgomery modulus defined for Scalar arithmetics, R^2 = (2^260)^2 % L

RR_FIELD
Montgomery modulus defined for FieldElement arithmetics, R^2 = (2^260)^2 % FIELD_L