Struct zerocaf::backend::u64::field::FieldElement

pub struct FieldElement(pub [u64; 5]);

A FieldElement represents an element of the field which has order of 2^252 + 27742317777372353535851937790883648493

In the 64-bit backend implementation, the FieldElement is represented in radix 2^52

Methods

impl FieldElement

pub const fn zero() -> FieldElement

Construct zero.

pub const fn one() -> FieldElement

Construct one.

pub const fn minus_one() -> FieldElement

Construct -1 (mod l).

pub fn is_even(self) -> bool

Evaluate if a FieldElement is even or not.

pub fn is_positive(&self) -> Choice

Checks if a ´FieldElement` is considered negative following the Decaf paper criteria.

The criteria says: Non-negative field elements. Let p > 2 be prime. Define a residue x ∈ F =Z/pZ to be “non-negative” if the least absolute residue for x is in [0,(p−1)/2], and “negative” otherwise.

Returns:

Choice(1) if pos.

Choice(0) if neg.

pub fn from_bytes(bytes: &[u8; 32]) -> FieldElement

Load a FieldElement from the low 253b bits of a 256-bit input. So Little Endian representation in bytes of a FieldElement.

pub fn to_bytes(self) -> [u8; 32]

Serialize this FieldElement to a 32-byte array. The encoding is canonical.

pub fn two_pow_k(exp: u64) -> FieldElement

Given a k: u64, compute 2^k giving the resulting result as a FieldElement.

See that the input must be between the range => 0..253.

NOTE: This function implements an assert! statement that checks the correctness of the exponent provided as param.

pub fn half_without_mod(self) -> FieldElement

Returns the half of an EVEN FieldElement.

This function performs almost 4x faster than the Half implementation but SHOULD be used carefully.

Panics

When the FieldElement provided is not even.

pub fn legendre_symbol(&self) -> Choice

Given a FieldElement, this function evaluates if it is a quadratic residue (mod l).

See: https://en.wikipedia.org/wiki/Legendre_symbol.

Returns:

-1 -> Non-quadratic residue (mod l) == Choice(0).

1 -> Quadratic residue (mod l) == Choice(1).

0 -> Input (mod l) == 0. Not implemented since you can't pass an input which is multiple of FIELD_L.

pub fn inverse(&self) -> FieldElement

Compute a^-1 (mod l) using the the Savas & Koç modular inverse algorithm. It's an optimization of the Kalinski modular inversion algorithm that extends the Binary GCD algorithm to perform the modular inverse operation.

The PhaseII it's substituded by 1 or 2 Montgomery Multiplications, what makes the second part compute in almost ConstTime.

Panics

It is not possible to invert 0 by obvious reasons. So an the function panics when trying to invert zero.

Special issue on Montgomery arithmetic. Montgomery inversion - Erkay Sava ̧s & Çetin Kaya Koç J Cryptogr Eng (2018) 8:201–210 https://doi.org/10.1007/s13389-017-0161-x.