Struct zerocaf::backend::u64::scalar::Scalar

pub struct Scalar(pub [u64; 5]);

The Scalar struct represents an Scalar over the modulo 2^249 + 14490550575682688738086195780655237219 as 5 52-bit limbs represented in radix 2^52.

Methods

impl Scalar

pub const fn zero() -> Scalar

Return a Scalar with value = 0.

pub const fn one() -> Scalar

Return a Scalar with value = 1.

pub const fn minus_one() -> Scalar

Return a Scalar with value = -1 (mod l).

pub fn is_even(self) -> bool

Evaluate if a Scalar is even or not.

pub fn into_bits(&self) -> [u8; 256]

Returns the bit representation of the given Scalar as an array of 256 bits represented as u8.

pub fn compute_NAF(&self) -> [i8; 256]

Compute the Non-Adjacent Form of a given Scalar.

pub fn compute_window_NAF(&self, width: u8) -> [i8; 256]

Compute the Windowed-Non-Adjacent Form of a given Scalar.

Inputs

width => Represents the window-width i.e. width = 2^width.

pub fn mod_2_pow_k(&self, k: u8) -> u8

Compute the result from Scalar (mod 2^k).

Panics

If the given k is > 32 (5 bits) as the value gets greater than the limb.

pub fn mods_2_pow_k(&self, w: u8) -> i8

Compute the result from Scalar (mods k).

Panics

If the given k > 32 (5 bits) || k == 0 as the value gets greater than the limb.

pub fn from_bytes(bytes: &[u8; 32]) -> Scalar

Unpack a 32 byte / 256 bit Scalar into 5 52-bit limbs.

pub fn from_bytes_wide(_bytes: &[u8; 64]) -> Scalar

Reduce a 64 byte / 512 bit scalar mod l

pub fn to_bytes(&self) -> [u8; 32]

Pack the limbs of this Scalar into 32 bytes

pub fn two_pow_k(exp: u64) -> Scalar

Given a k: u64, compute 2^k giving the resulting result as a Scalar.

See that the input must be between the range => 0..250.

Panics

If the input is greater than the Sub-group order.

pub fn half_without_mod(self) -> Scalar

Returns the half of an EVEN Scalar.

This function performs almost 4x faster than the Half implementation but SHOULD be used carefully.

Panics

When the Scalar provided is not even.